requirements.txt
asgiref==3.7.2
Django==5.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
dnspython==2.6.1
gunicorn==21.2.0
mysqlclient==2.2.1
packaging==23.2
pymongo==4.6.2
PyMySQL==1.1.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
whitenoise==6.6.0
